backend/storyboard-agent/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
numpy
supabase
langchain-core
langchain-teddynote
